+++ tests/doc_examples/snappylapy/fixtures/test_fixture_expect_dataframe.py
@@ -1,12 +1,16 @@
 import pytest
-try:
-    import pandas as pd  # noqa: F401
-    pandas_installed = True
-except ImportError:
-    pandas_installed = False
 from snappylapy.fixtures import Expect
 
-@pytest.mark.skipif(not pandas_installed, reason="pandas is not installed")
+try:
+    import pandas as pd
+
+    HAS_PANDAS = True
+except ImportError:
+    HAS_PANDAS = False
+
+
+@pytest.mark.skipif(not HAS_PANDAS, reason="pandas is not installed")
 def test_expect_dataframe(expect: Expect) -> None:
+    """Test that dataframe matches the snapshot."""
     df: pd.DataFrame = pd.DataFrame({"key": ["value1", "value2"]})
     expect.dataframe(df).to_match_snapshot()